+++ test/js/jsx/index.jsx
@@ -1,5 +1,5 @@
 import { append, pollQuerySelector } from 'a-b-doer';
-import { Simple, RefHook, Hooks, Switch } from './templates';
+import { Simple, RefHook, Hooks, Switch, OrderApp } from './templates';
 import './styles.scss';
 
 pollQuerySelector('html body', (body) => {
@@ -12,4 +12,12 @@ pollQuerySelector('html body', (body) => {
 	append(<Switch id="tpl7" />, body);
 	const Title = document.querySelector('h1');
 	append(<Title id="tpl8" data-o="h1" class="title" />, body);
+	append(
+		<OrderApp id="tpl9">
+			<Simple class="simple-sub" id="tpl9-1" />
+			<Simple class="simple-sub" id="tpl9-2" />
+			<Simple class="simple-sub" id="tpl9-3" />
+		</OrderApp>,
+		body
+	);
 });